Work on app.R to expose prior therapy
Please DO NOT MERGE to master yet, as this PR is only intended as a vehicle to ask questions.
This small change just queries the new DB table for prior therapy info, and blindly displays it next to the corresponding trial, regardless of whether any of the associated thesaurus terms match the patient's prior therapies.  I'm submitting it for review at this stage since I suspect that I might be missing something, even at this simple level.

There are TODOs in the code showing where I need to call the new check_if_any() function that will implement the logic we discussed in the last meeting.